renamed gomoku.java to consolegomoku.java
